Set the viewport before navigating to the site.
The docs of page.setViewport function say: "A lot of websites don't expect phones to change size, so you should set the viewport before navigating to the page".

We were setting the view port after navigating to the page and this was causing layout issues.
